edge_server/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
vaderSentiment
pydantic
requests
python-dateutil
